refactor: [3] add test for get connection of mongo
